modified deliveries/notifiers structure
