--- spec/notifiers/telegram/company/long_time_review_report_payload_spec.rb
@@ -1,25 +1,0 @@
-# frozen_string_literal: true
-
-describe Telegram::Company::LongTimeReviewReportPayload, type: :service do
-  subject(:service_call) { described_class.new.call(insightable: insightable) }
-
-  let!(:insightable) { create :company }
-
-  context 'without pull requests' do
-    it 'generates payload' do
-      expect(service_call).to eq "Company #{insightable.title} doesn't have long time review"
-    end
-  end
-
-  context 'with pull requests' do
-    let!(:repository) { create :repository, company: insightable }
-
-    before { create :pull_request, repository: repository, created_at: 3.weeks.ago }
-
-    it 'generates payload' do
-      expect(
-        service_call.include?("Repository long time review of #{repository.title}")
-      ).to be_truthy
-    end
-  end
-end
